feat: #1325 Update webhook edit + create modal tests
